+++ src/actions/delete-maintenance.ts
@@ -1,6 +1,9 @@
 'use server'
 
+import { subject } from '@casl/ability'
+import { defineAbilityFor, type Subjects } from '@inspetor/casl/ability'
 import { prisma } from '@inspetor/lib/prisma'
+import type { AuthUser } from '@inspetor/types/auth'
 import { returnsDefaultActionMessage } from '@inspetor/utils/returns-default-action-message'
 import z from 'zod'
 
@@ -13,7 +16,29 @@ export const deleteMaintenanceAction = authProcedure
       dailyMaintenanceId: z.string(),
     }),
   )
-  .handler(async ({ input }) => {
+  .handler(async ({ input, ctx }) => {
+    const dailyMaintenance = await prisma.dailyMaintenance.findUnique({
+      where: { id: input.dailyMaintenanceId },
+    })
+
+    if (!dailyMaintenance) {
+      return returnsDefaultActionMessage({
+        message: 'Manutenção diária não encontrada',
+        success: false,
+      })
+    }
+
+    const ability = defineAbilityFor(ctx.user as AuthUser)
+    const subjectMaintenance = subject('MaintenanceDaily', {
+      companyId: dailyMaintenance.companyId,
+    }) as unknown as Subjects
+    if (!ability.can('delete', subjectMaintenance)) {
+      return returnsDefaultActionMessage({
+        message: 'Sem permissão para excluir manutenção diária',
+        success: false,
+      })
+    }
+
     await prisma.dailyMaintenance.delete({
       where: {
         id: input.dailyMaintenanceId,
